freq/project_2/indicators/tv/v2/requirements.txt
python-binance>=1.0.16
python-dotenv>=1.0.0
rich>=13.0.0 